Remove encryption to crypto package
